Time: 0 ms (100.00%) | Memory: 8 MB (41.33%) - LeetSync

+++ 476-number-complement/number-complement.cpp
@@ -4,7 +4,7 @@ public:
        int n=num;
        int c=0;
        long long r=0;
-       long long power=1;
+       long  power=1;
        while(n!=0)
        {
            c=n%2;
